fixup! fix: warning when tries to start an empty workspace
